+++ src/plugins/Morse.cpp
@@ -31,7 +31,7 @@ namespace mesmer
     virtual void canPrtnFnCntrb(gDensityOfStates* gdos, double beta, double &PrtnFn, double &IntrlEne, double &varEne) ;
 
     // Function to return the number of degrees of freedom associated with this count.
-    virtual unsigned int NoDegOfFreedom(gDensityOfStates* gdos) ;
+    virtual size_t NoDegOfFreedom(gDensityOfStates* gdos) ;
 
     // Provide a function to calculate the zero point energy of a molecule.
 	  virtual double ZeroPointEnergy(gDensityOfStates* gdos) ;
@@ -181,7 +181,7 @@ namespace mesmer
   }
 
   // Function to return the number of degrees of freedom associated with this count.
-  unsigned int Morse::NoDegOfFreedom(gDensityOfStates* gdos) {
+  size_t Morse::NoDegOfFreedom(gDensityOfStates* gdos) {
     return m_vibFreq.size() ;
   }
